Redesigned SensitiveDetector for PXD/SVD
Since the SensitiveDetector implementations of PXD and SVD are almost
identical we combined them to one template class which can create both
collections as neccessary.

This means that the constructors for {PXD,SVD}{Sim,True}Hit have to be
compatible to allow templating. But if required, different
specialisations of the base class could be defined for PXD and SVD.

Furthermore, the TrueHit generation has been modified: It should now
work for PXD and SVD and will create TrueHits which mark the traversal
of the detector plane (z=0) and contain informations about the momentum
of the particle when entering and when exiting the silicon

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@2021 b38c6c8f-bee8-465d-8448-9abe6915d496