Added MaterialScan Module
Added a new module which allows to determine the Material budget of the
geometry in different ways:

- Spherical Scan: Shoot rays from the origin and scan theta and phi. The
  range for theta, phi and the number of rays can be defined using the
  module parameters
- Planar Scan: Shoot rays perpendiular to a given plane. The plane,
  range and number of rays can be defined freely using the module
  parameters.

For both types of scans the depth of the scan can be limited to provide
cut views through the detector at any plane.

The output will be a root file containing 2D histograms of the scanned
X0 and lambda per defined G4Region.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@2077 b38c6c8f-bee8-465d-8448-9abe6915d496