Added entry and exit coordinates to TrueHits
This is a small change to add the entry and exit coordinates to the
VXDTrueHits (and the PXD,SVD specializations). This patch only stores
the u,v projections so if the particle enters the sensitive area from
the side the w coordinate is lost.

Sadly the diff is very large as there have been some files with dos line
endings which now got converted to unix style.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@2657 b38c6c8f-bee8-465d-8448-9abe6915d496