Re-add range check to StoreArray::operator[], add nextFreeAddress() for use with placement-new.
Removing the range check in r3261 broke the TOPReconstruction module, sorry for that.

Since there is no safe way to have operator[] do a range check and be compatible
with placement-new at the same time, I added nextFreeAddress() to provide the same
functionality as ->AddrAt(array.getEntries()), but with no possibility of creating
a non-contiguous array.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@3322 b38c6c8f-bee8-465d-8448-9abe6915d496