Improve naming conventions in RelationIndex
The following functions have been renamed:
getFrom(FROM*)    -> getElementsFrom(FROM*)
getFirstTo(FROM*) -> getFirstElementFrom(FROM*)
getTo(TO*)        -> getElementsTo(TO*)
getFirstFrom(TO*) -> getFirstElementTo(TO*)

This makes it obvious that objects of type RelationIndex::Element are returned
(or ranges of them), and which side of the relation you want to match with the
given pointer.

The original functions will be removed once they're no longer used.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@3442 b38c6c8f-bee8-465d-8448-9abe6915d496