The CDCHit is now to my best knowledge prepared for the Output of the CDCUnpacker, that is to be written in the near future. It contains not only the drift time (format is still unknown, asked Nanae-san to help with this), but as well the ADC count (instead of the charge directly), and some core, which can be used together with some new WireID object.
The new WireID object handles the different encoding of wire number in a consistent way. CDCGeometryPar and CDCSimHit are updated to reflect this.

The CDCRecoHit can't take the information anymore directly from the CDCHit. Instead, there are new objects, so called "translators".
The CDCRecoHit object now has 4 new static members. This are 3 pointers to translator classes, that must be set before using the CDCRecoHits.
In addition there is an update switch. If the switch is set to false, the CDCRecoHit will just the values for position and drift length,
that it had at construction time.
If the switch is set to true, it will try to update during each call of the newly implemented getMeasurement function. This 
function is called from Genfit during the fitting. So far it took directly the inherited function from AbsRecoHit and still does,
if the switch is false. However, if the switch is set to true, it will first update some stuff - or not, as this part isn't yet implemented.
This mechanism allows to use information only accessible during the tracking to be used for the determination of the values, e.g.
for the translation of drift time to drift length the z position is needed to take the propagation in wire effect into account.

The translators are now only the simplest form of translators, taking the ideal geometry and otherwise try to mirror the current simple
digitization.

In addition the CDCGeometryPar now gives back a reference instead of a pointer in its Instance method.

Furthermore various doxygen comments have been updated or were created for the first time.


git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@3583 b38c6c8f-bee8-465d-8448-9abe6915d496