Handing a pointer to Null pointer to SetBranchAddress() is bad
this creates a new object owned by us at the given position, which for some
reason I don't entirely understand is not only a memory leak, but also results
in branches not being written, even if they have ptr != NULL when TTree::Fill()
is called.

Instead we should either give it a valid TObject**, or NULL (which creates
a default object, but owned by the data store).
Fixes #715

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@3585 b38c6c8f-bee8-465d-8448-9abe6915d496