Added possibility to use multiple Generators
Up to now the MCParticleGraph was always removing the existing particles
when filling the MCParticle list to ensure consistency of the whole
list. Now the existing particles are only removed when the option
MCParticleGraph::c_clearParticles is specified when generating the list,
for example after simulation, otherwise the particles are just appended
to the list of already existing particles

This might break the ordering of particles in the MCParticle list (up to
now it was guaranteed that the particles are ordered by depth, primary
particles first, then second generation and so forth) so if some code
depends on this ordering it might not work anymore.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@3819 b38c6c8f-bee8-465d-8448-9abe6915d496