Fix a Bug in getIndex() of MCParticle
When the RootInput Module, the index returned by MCParticle::getIndex()
and MCParticle::getArrayIndex() was not always correct. The member
storing the index is transient to save space but the calculation of the
index was not triggered correctly

The getMother()/getDaughter() routines and Relations were not affected
by this bug, the only known problem so far was in the PrintMCParticles
module.

Also, removed the obsolete compatibility definition for
DEFAULT_MCPARTICLES

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@3944 b38c6c8f-bee8-465d-8448-9abe6915d496