Added module to remove MCParticles based on cuts
The RemoveMCParticles module can be used to perform cuts on the list of
the MCParticles. Possible cuts are:
 - z vertex position
 - rphi vertex position
 - transverse momentum
 - theta of the flight direction

BEWARE: Any existing Relations to that MCParticle collection beomce
invalid and are not fixed. This module is intended to be used directly
after generation to e.g. perform acceptance cuts or similar.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@3947 b38c6c8f-bee8-465d-8448-9abe6915d496