Make Const::ParticleType iterable (see Const::ParticleSet class documentation for examples)
In addition to the chargedStable set, there's now also a corresponding ChargedStable class
deriving from ParticleType that can be used e.g. as a function argument to allow only charged
stable particles to be passed. Const::ChargedStable::electron, muon, etc. are all part of the global
chargedStable set and so provide a fixed index() from 0 to 4.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@4637 b38c6c8f-bee8-465d-8448-9abe6915d496