Extend RelationArray::consolidate() (by Martin Ritter)
consolidate() now expects a new index and a bool flag for each element.
The bool should be true if the element got attributed to another element
(e.g. the particle got ignored, energy should be attributed to it's
mother) or was just reordered (i.e. TrackID != particle index).

In addition, there is a new parameter "action" which determines the
behaviour for elements which got reattributed. The default is
c_doNothing which is the same as before and treats reattribution and
reordering the same way. In addition there are c_zeroWeight which will
set the weight to zero if the element got reattributed, c_negativeWeight
which will ensure the weight is negative for reattributed elements and
c_deleteElement which will drop all reattributed elements

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@5598 b38c6c8f-bee8-465d-8448-9abe6915d496