Added PXD + SVD DockBoxes to Simulation
This commit adds the DockBoxes for the PXD+SVD to the simulation. Each
Box is about 4.5kg, mainly Aluminum and there are 10 boxes on backwards
side and 8 boxes on forward side just outside the CDC.

Material Description might change in the future as design of those boxes
is still in flux but placement and dimensions are considered final.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@6143 b38c6c8f-bee8-465d-8448-9abe6915d496