In the default simulation setting, the secondary particle information is not stored in the final MCParticles block except some specified cases. If one wants to store all the secondary particle information, one should set the simulation parameter "StoreAllSecondaries" as true.
It has been noted that when the "StoreAllSecondaries" is set to true, the number of stored secondary particles is too high, since there are a huge number of very low energy photons, eletrons, and etc. Usually, we do not need information from such particles. Therefore, Marko Staric suggested to revive the "SecondariesEnergyCut" parameter. The default value for this parameter is 1MeV. You can adjust the cut by giving it a different number. The information of secondary particle with energy less than this cut will not be stored in the MCparticles block, even if "StoreAllSecondaries" is set to true. 

Note: There are some special classes of secondary particles mentioned in the first paragraph. They could be decays-in-flight or particles tagged as unignored by a subdetector group. The information from these particles will be always stored in the "MCParticles" block regardless of the "StoreAllSecondaries" setting.



git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@6382 b38c6c8f-bee8-465d-8448-9abe6915d496