Add RandomBarrier module to allow multiple TRandom3 instances
E.g. a module chain of the sort [ParticleGun -> RandomBarrier ->
FullSim] would use one RNG instance for the ParticleGun, and another
for FullSim and all following modules. You may find this useful if you
want to change the simulation, but don't want differences to affect
the particle generation.
The output is equivalent to saving the output of ParticleGun in a file,
and reading it again to do the simulation.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@6404 b38c6c8f-bee8-465d-8448-9abe6915d496