Add possibility to change magnetic field stepper
Geant4 has a multitude of different steppers for different magnetic
fields. Currently we use the ClassicRK4 Stepper which is a 4th order
Runge Kutta integrator.

According to the Geant4 manual[1] there might be others more suited
depending on our magnetic field. Especially the Nystrom stepper which
"provides accuracy near that of G4ClassicalRK4 (4th order) with a
significantly reduced cost in field evaluation" or the more simple
HelixExplicitEuler or HelixSimpleRunge tracker optimized for pure
magnetic fields (no electrical component)

This patch allows to choose the Stepper used in the simulation as a
parameter to the FullSim module. All other values for propagation in
magnetic fields are left to the default values (MinimumEpsilon,
MaximumEpsilon, DeltaOneStep, DeltaIntersection, MinStepSize,
MaxStepSize).

The possible values for the magneticField parameter are:
 - none: Use no Field at all. Much faster than setting the field to 0
   but the CDC sensitive detector assumes a field and thus crashes with
   this setting
 - default: Use the default ClassicRK4 Stepper
 - nystrom: Use the new NystromRK4 Stepper
 - expliciteuler: Use the HelixExplicitOrder Stepper (1st order)
 - simplerunge: Use the HelixSimpleRunge Stepper (2nd order)

In addition, the magnetic field can be wrapped in a Cache: If the
geometrical distance between two bfield lookups is less than a given
minimum the first value will be returned again. The default setting is
not to use caching.

[1] http://geant4.web.cern.ch/geant4/UserDocumentation/UsersGuides/ForApplicationDeveloper/html/ch04s03.html

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@6496 b38c6c8f-bee8-465d-8448-9abe6915d496