Optimized PXD Clustering
some restructuring of the PXD clustering to speed up some part and
remove redundant code.

There are a lot of changes in this code but as most of it happend
incremental it was not feasible to split it to smaller chunks

Pixel: The pixel class we used in clustering kept a pointer to the
PXDDigit and obtained the values (u, v, charge) directly from there.
As is needed to work without a valid pointer this caused some
redirection and uncessary ugly code. Now the pixel just remembers the
index of the PXDDigit and copies the needed values.

ClusterCandidate: each ClusterCandidate now also has a Pointer to the
parent cluster in case it got merged into another cluster. This allows
to get rid of the broadcasting in the ClusterCache which was needed to
update the Cluster pointers upon merging. To prevent relocations the
vectors to contain pixels get a default capacity>0 now.

ClusterCache: The cache now keeps the list of clusters itself and will
always return a reference to a ClusterCandidate when findCluster(x,y) is
called. One just needs to add the Pixel to that ClusterCandidate. To
prevent allocations and free calls the ClusterCache will keep the list
of Clusters between events and just reuse it.

ClusterProjection: New class to aggregate information about the cluster
projected on one axis, needed for Position and Error calculation. This
removes some convinient but ugly macros

PXDClusterizer: Calculation of Position and Errors has been delegated
into a seperate function and is now the same code for each axis,
removing the possibility of U/V typos.

Also, most of the time of the clustering was apparently spent in
creation of the RelationIndex objects. As we know much more about the
Digits->X relations than RelationIndex assumes we can do much better: We
know that the relations are created in an optimal way by the Digitizer
so we can exploit the fact and make our own lookup table by using an
vector where the nth element is a pointer to the RelationElement of the
nth Digit (or null). At the worst case scenario of 3% occupancy this
would be 2*768*250*40*0.03*sizeof(T*) which is around 4MB for both lookup
tables combined and allows us to double the speed of the Clusterizer for
simulated data.

Alltogether, tests suggest a speed improvement of more than factor 2 for
simulated data (Particle Gun) by keeping all the functionality

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@6497 b38c6c8f-bee8-465d-8448-9abe6915d496