Add PXDDigit sorting Module
This module allows to sort an existing PXDDigit collection as the
clusterizer needs the digits in sorted order.

Readout and simulation produce the digits in a sorted way by default but
in special cases this might be needed. Up to now, the Clusterizer was able
to sort the digits but as it is almost never used we put it into a
separate module.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@6500 b38c6c8f-bee8-465d-8448-9abe6915d496