Add optional geometry components
This change adds the possibility for optional geometry components. If
the component is defined with the attribute isDefault="false" in the xml
file it will not be created by default but it can be specified in the
list of components

In addition, there is a new parameter to the geometry module called
"additionalComponents". This is a list of components which should be
created in addition to the default set. If a specific list of components
is provided using the "components" parameter than this parameter is
ignored (and produces a warning)

The main use case is to allow easily changing the magnetic field. To
replace the default magnetic field with the 3D magnetic field one can
use

geometry.param({
    "excludedComponents": ["MagneticField"],
    "additionalComponents": ["MagneticField3d"],
})

Of course, the magnetic field can also be specified directly in the list
of components, example to only simulate the tracking components with a
constant magnetic field limited to the CDC

geometry.param("components", ["Beampipe","PXD","SVD","CDC","MagneticFieldConstant4LimitedRCDC"])

The currently defined fields are
- MagneticField: constant 1.5T in the whole detector, default
- MagneticField2d: field from 2D magnetic field map
- MagneticField3d: field from 3D magnetic field map
- MagneticField4LimitedRSVD: constant 1.5T for R<155mm, |z|<4m, zero
otherwise
- MagneticField4LimitedRCDC: constant 1.5T for R<1137.2mm, |z|<4m, zero
otherwise

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@8076 b38c6c8f-bee8-465d-8448-9abe6915d496