ParticleGun: Add more distributions
This is a fairly large update to add new distributions
 - normalCos: normal distributed cosine of an angle
 - inversePt: uniform in track curvature
 - polyline: arbitary distribution following 1D polyline PDF
 - polylinePt: same as polyline but for the transverse component
 - polylineCos: same as polyline but for the cosine

The uniformCosinus and discreteSpectrum distributions have been renamed
to "uniformCos" and "discrete". A Warning is issued if the old names are
used.

In addition, the sanity checks on all parameters have been
reimplemented and a unittest for these checks has been added.

The examples have been updated to reflect the current version.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@8111 b38c6c8f-bee8-465d-8448-9abe6915d496