Extend display module to make use of trajectories
the DisplayModule will now use the MCParticleTrajectories to
draw the MCParticle tracks if possible. The position/momentum from the
simulation will be used as reference points in TEveTrack to make sure
the track is always at the right position.

Between the references, TEveTrack will propagate the track on its own to
create smooth track representation even for coarse Geant4 stepping

To make use of this feature, the "trajectoryStore" parameter of the
FullSim module has to be set to at least 1, preferrably 2 or 3

fullsim.param("trajectoryStore", 2)

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@8119 b38c6c8f-bee8-465d-8448-9abe6915d496