Allow easy profiling of single modules
Instead of creating separate data files for initialize, beginRun, event,
endRun and terminate we now only profile one module at a time. If a
module name was set for profiling the callgrind instrumentation is
enabled just for calls to that module.

This adds a new parameter "profile" to basf2 if callgrind is available.
This parameter allows to specify the name of a module which should be
profiled using callgrind. If valgrind is not already running it is
started automatically and the profile data is written to

callgrind.<ModuleName>.<PID>

Multiprocessing is disabled if the profile option is present.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@8472 b38c6c8f-bee8-465d-8448-9abe6915d496