Simple run-dependency for Gearbox on file level
Allow to have some run dependency in xml files by replacing {EXP} and
{RUN} in the file backend uri with the actual experiment and run number
respectively. Optionally, the number of leading digits can be specified
separated by a colon, e.g. {EXP:4} to have the experiment number
starting as 0001.  In addition, if the file cannot be found the
directory separators "/" will be successivley replaced by "-" to allow
for flat file structure.

To use this feature, use something like

>> gearbox.param("backends", ["file:/rundata-{EXP:3}-{RUN:4}/","file:"])

in the steering file to search for run-dependent files for experiment 3,
run 1 in  in rundata-003-0001/ and fall back on the default files if no
run-dependent version exists. It is also possible to supply only one,
either {EXP} or {RUN} in the backend string.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@10856 b38c6c8f-bee8-465d-8448-9abe6915d496