Reimplement VXD SimHits and SensitiveDetector
WARNING: This commit breaks compatibility with older files containing
PXDSimHit, SVDSimHit, PXDTrueHit or SVDTrueHit objects.

This is a very large commit to reimplement the VXD SensitiveDetector
implementation and the SimHits for PXD and SVD. Aim is to reduce disk
space when storing sim hits by retaining the required precision. As all
files depend on each other this is sadly one large commit.

Previously, we saved a SimHit for each Geant4 step if enough energy was
deposited. Now we collect all steps and calculate the minimal number of
SimHits so that all step points are within a given tolerance to the linear
approximation for each SimHit using the Ramer-Douglas-Peucker algorithm.

In addition, the energy deposition profile for all steps belonging to a
SimHit is simplified using the same algorithm. The points of the profile
are then saved using an optimized encoding using 10bits for the position
and 22bits for the number of electonrs. VXDSimHit has member functions
to sample this optimized energy deposition profile and PXDSimHit and
SVDSimHit now inherit from VXDSimHit.

The tolerances for SimHit creation and energy deposition can be
specified separately for PXD and SVD in the XML files.

see https://indico.mpp.mpg.de/getFile.py/access?contribId=30&sessionId=1&resId=0&materialId=slides&confId=2809

Most of the heavy lifting is now in SensitiveDetectorBase as it did not
depend on the templates anyway. Just the creation of objects is
delegated to the templated class.

We expect a file size reduction of the SVD/PXD objects by factor 20 to
30 depending on chosen tolerances.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@10859 b38c6c8f-bee8-465d-8448-9abe6915d496