Initialize PhyiscsList tables on initialize
So far, the lookup tables and other stuff needed by the Geant4
PhyiscsList were calculated on beginRun(). As these structures are
readonly this is a waste of memory and cpu time when running in parallel
processing mode.

As Geant4 does a thorough clean up on exit which modifies the geometry
we have to jump through some loops to prevent memory blow up due to page
duplication.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@11390 b38c6c8f-bee8-465d-8448-9abe6915d496