Add tool to measure memory usage of the framework
framework/examples/tools/memoryusage monitors the memory usage of a
given program and creates a memory over time graph. It follows all child
processes (e.g. basf2 multi processing) and takes page sharing into
account. If the log level of the basf2 framework is set to
info it can also determine the role of the child processes from the
output of the ProcManager and label them accordingly.

Only works on Linux and requires matplotlib to work

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@11391 b38c6c8f-bee8-465d-8448-9abe6915d496