Fix bug in VXD digit/cluster relations
The negative sign of relations from secondaries which were not stored
got lost in digitizer and clusterizer. Now all negative relations are
ignored and not propagated to digits/clusters

Also, the {PXD,SVD}TrueHit relations are deleted when the particle is
ignored (instead of setting it to negative weight). The simhit relations
are still kept with negative sign for now but could be removed as well.
There should be no difference on digit/cluster level.

closes #1353

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@15819 b38c6c8f-bee8-465d-8448-9abe6915d496