Add initial PXD/SVD configuration for Beast phase2
This commit add pxd/data/PXD-phase2.xml and svd/data/SVD-phase2.xml
which are copies of the normal configuration (pxd/data/PXD.xml and
svd/data/SVD.xml respectively)

For PXD the only change is removal of most of the ladders and the
envelope volume.

For SVD the changes are are  larger:
 - remove envelope volume
 - remove most ladders
 - put ladder 1 in each layer at phi=0
 - remove ladder shift to center sensors at x=0
 - remove svd support cones and endrings
 - TODO: add cooling pipes

Most of these changes are done by only including portions of the
original support and component definitions so most changes in phase3 geometry
should propagate to phase2 automatically.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@16948 b38c6c8f-bee8-465d-8448-9abe6915d496