Add DataStore info panel to display
Provides overview of DataStore contents, object details,
and relations. Navigation is coupled to display selection,
i.e. selecting an object in the display shows info about it,
and opening an object's info page selects it in the display
(if it has a visualisation).
Also contains stand-alone inspector.py example

Other changes:
Make VisualRepMap a singleton (it connects objects in DataStore and Eve, both of which have global state anyway.)
Hide VisualRepMap implementation

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@17724 b38c6c8f-bee8-465d-8448-9abe6915d496