Revert "make FileSystem a namespace instead of class"
for the 2nd time.
The reason behind this are problems of root6,
where functions in namespaces are not found unless you explicitly include
the header file, or use ROOT.gInterpreter.ProcessLine('#include "myheader.h"')
in python. The developers are aware of the problem, but are at the 'plan
to have a meeting' stage, so this might take a while.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@19469 b38c6c8f-bee8-465d-8448-9abe6915d496