Add mdst compatibility check
added a steering file to print the full contents of mdst dataobjects (and
EventMetaData) to the standard output and the expected output for a 10
events v00-05-02 mdst file.

This allows to check the compatibility of mdst dataobjects as the output
will change if there is a significant change to the mdst classes.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@19530 b38c6c8f-bee8-465d-8448-9abe6915d496