Use new RandomGenerator for the framework
this commit touches a lot of framework parts to use the new
RandomGenerator during event processing.

It will create two instances of the RandomGenerator, one is event
dependent for normal event processing and one which is independent for
begin/end run processing. The event dependent one is passed to the next
process in multi processing to make sure that the random number state
does not depend on single/multi processing.

In begin/end run the second generator will be seeded just from
experiment and run number. It will have the same state in all processes
so consuming random numbers in beginRun()/endRun() will be identical in
all processes.

The format of FileMetaData and set_random_seed() is changed as the new
random generator takes an arbitary string as seed. The TRandom3 instance
is removed from FileMetaData.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@20035 b38c6c8f-bee8-465d-8448-9abe6915d496