MCInitialParticles/BeamParameters: return idendity transformation when generating in CMS
When generating in CMS the nominal beam parameters are still stored in
lab system to be able to do the smearing but the generated events are
actually in CMS. This leads to inconsistencies when calculating lab/cms
values as the boost is wrong.

With this fix the BeamParameters will return the identity transformation
when c_generateCMS is set, this should at least ensure consistency with
the analysis tools.

Also, c_generateCMS is now always included in the allowed options for
event generation and cannot be disabled.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@20278 b38c6c8f-bee8-465d-8448-9abe6915d496