Fix bug when a correlated electronic noise was miscalculated:
for (int T_clock = 0; T_clock < 31; T_clock++) {
   AdcNoise[T_clock] = 0;
   for (int iCal = 0; iCal < T_clock; iCal++) { // <= wrong conditon
      AdcNoise[T_clock] = m_elecNoiseM[T_clock][iCal] * genNoise[iCal] + AdcNoise[T_clock];
   }
}

In the inner loop condition iCal < T_clock cause that the diagonal elements of the matrix m_elecNoiseM
are never used for noise calculation so calculated noise was 50% less than expected from
the covariance matrix.

The function generateCorrelatedNoise is introduced in ECLNoiseData class to avoid the large
matrix export outside the class.

Also some optimization in ECLWaveformData.h to get rid of huge arrays inside the header file
where they are not really needed since data are placed in the specific order which can be exploited.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@20395 b38c6c8f-bee8-465d-8448-9abe6915d496