Fix gcc 5.2 compile error
Apparently you can neither compare to stringstreams nor stream one
stream into the other, that doesn't seem to be defined in the standard.
For some reason it worked so far but the new version of the libstdc++
in gcc 5 seems to have removed this. Luckily there were not many cases.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@21862 b38c6c8f-bee8-465d-8448-9abe6915d496