Create calibration_framework script and update calibration example to use calibration_framework
calibration_framework provides two functions: add_calibration() and
run_calibration()

add_calibration() needs to be called with two arguments: collector module
and algorithm (or list of algorithms). It will set the prefix correctly
and setup the structures needed to run the calibration

run_calibration() needs to be called with a path prior to the
calibration and optionally an directory to store the results, a maximum
number of iterations and a callback function to be called before each
iteration (mainly to allow for modification of the random seed).

It will run the calibration and iterate if needed. The output of all
calibrations will be stored in CalibrationResults.xml in the output
directory (default is current directory).

The state of the calibration will be independent between iterations: you
cannot see any changes in member variables between iterations. In each
iteration any payloads created by previous iterations will be visible
and take precedence over payloads from the database setup prior to
calling run_calibration().

It shoult be multiprocessing save as all the algorithms will be called
once processing is finished.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@22182 b38c6c8f-bee8-465d-8448-9abe6915d496