scons: Check consistency of externals version
when the default externals in .externals changes, setuprel has to be
called again to actually use them. As calling setuprel after svn up can
be easily forgotten we now check if the versions are consistent: if a
.externals exists and BELLE2_EXTERNALS_VERSION is defined in the
environment we check that they actually are the same. If they differ we
stop.

To not perform this check one can add --ignore-externals-version as
argument to scons

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@22205 b38c6c8f-bee8-465d-8448-9abe6915d496