Add skipTillEvent parameter to EventInfoSetter and RootInput modules
This allows to skip to a specific event before processing. For example

   >>> eventinfosetter.param("skipTillEvent", [1,1,1])

will start processing at experiment=1, run=1, event=1.

For EventInfoSetter:

If no experiment,run (1,1) is to be generated then all events will be
skipped.  If the event number is larger than the number of events to be
generated for that run then generation will start with the first event
in the next run.

For RootInput:

If not such experiment,run,event exists in the files to be read an error
is printed and processing starts from the beginning.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@22451 b38c6c8f-bee8-465d-8448-9abe6915d496