Add initial implementation of python interface to ConditionsDB and upload script
This adds upload_localdatabase to the framework tools which allows a
local database to be uploaded to the ConditionsDB in one go

$ upload_localdatabase -j 8 --tag MyGlobalTag localdb/database.txt

will upload all payloads it finds in localdb/database.txt to the
ConditionsDB and create iovs in the global tag MyGlobalTag using 8
parallel connections.

As the ConditionsDB currently does not support unlimited iovs, -1 end
experiment and runs need to be converted. The argument --max-exp needs
to be supplied to indicate which experiment to use in case of -1 and
run=-1 will be set to the highest run defined in the experiment.

TODO: if the same payload is referenced more than once in the
database.txt it will be uploaded more than once. However the local
database backend does not support this anyway so this would only be
applicable to hand crafted database files.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@22452 b38c6c8f-bee8-465d-8448-9abe6915d496