Initial implementation of new BFieldManager
The aim is to move the existing BFieldManager to framework so that it's
accesible from mdst objects without additional dependencies. The
implementations of the fields can still stay in geoemetry.

If no magnetic field is defined the first try to evaluate the magnetic
field will create an error message and a constant magnetic field with
1.5T along z is returned. This behaviour will change in the future.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@22593 b38c6c8f-bee8-465d-8448-9abe6915d496