Fix problems with MaterialScan
stuck tracks: Sometimes geantinos get stuck so nudge them along. If they
.still don't budge kill this one track, not the whole processing. In
principle this should be fixed by Geant4, we should make a bug report about this.

memory leak: there was a very obvious and very big memory leak in the
MaterialScan. Nobody ever saw it because usually MaterialScan doesn't
run that long so it's not a problem but with many step points it becomes
an issue.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@22796 b38c6c8f-bee8-465d-8448-9abe6915d496