Change genfit Geant4 material interface
Up to now the Genfit <-> Geant4 interface clamped the position of tracks
to be inside the world volume due to problems with Geant4.9. However
this could cause trouble at the borders.

The new approach is to virtually extend the world volume to infinity: If
the current track position is outside the world volume just return
the distance to the world volume as possible step length and assume the same
material as in the world volume.

This makes the code slightly faster as lookup outside of the world
volume is much simplified and it should also get rid of the Geant4
exceptions we now see with Geant4.10.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@22803 b38c6c8f-bee8-465d-8448-9abe6915d496