Clear BFieldMap on geometry creation
so far the map was not cleared if the geometry was created multiple
times which lead to stronger magnetic field (1.5T, 3T, 4.5T and so
forth) depending on how often the geometry module was included

Note: multiple instances of the Geometry modules are still not a good
idea. The geometry will be destroyed and created but there are probably
memory leaks and if there are differen parameters for the different
modules the state is not well defined (the last geometry wins).

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@23129 b38c6c8f-bee8-465d-8448-9abe6915d496