Add the possibility to annotate the memoryusage profile
Allows to specify a regular expression to look for in the output of the
program. If the expression matches the line will be added to the
memoryprofile at the given time stamp and should allow to mark special
positions like "begin run" or "optimize geometry" in the plot if the
output is provided in suitable format.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@23661 b38c6c8f-bee8-465d-8448-9abe6915d496