Start documenting the python interface through boost python
this helps when using the help() function in python, the ? magic in
ipython or the pydoc3 utility

for example, run "pydoc3 basf2_def" to see a documented version of the
global functions in the basf2 module.

TODO: there are still a lot of undocumented boost::python exposed
functions and we should make sure that pydoc3 basf2 shows all and don't
require to look for basf2_def. Not yet sure how to do that though.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@23663 b38c6c8f-bee8-465d-8448-9abe6915d496