BeamParametersModule: change how existing beamparameters are handled
before there was an error when there was an existing BeamParameters
object. Now the default is to do nothing if the BeamParameters are
already present (from file or due to multiple modules) and just print a
warning.

There's a new parameter "overwrite", if set to true any existing
BeamParameters object will be overwritten.

Also, when using the add_beamparameters function we now add the name to
the preset to the module name shown in the statistics.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@24294 b38c6c8f-bee8-465d-8448-9abe6915d496