Move BeamParameters from datastore to dbstore
The BeamParameters are now stored in a DBObjPtr instead of a
StoreObjPtr. A default version on the Y4S is provided by the framework
so add_beamparameters() is now optional and only needs to be added if a
different setting should be used.

The "overwrite" parameter of the BeamParameters module has been removed
as it's no longer meaningful. Additionally the BeamParameters module can
now create BeamParameters objects for the database directly.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@25554 b38c6c8f-bee8-465d-8448-9abe6915d496