b2bii: use central database
added a function "setupB2BIIDatabase()" to b2bii/scripts and the
ConvertMdstModule will now check if the objects from the database are
actually what we expect. For MC it will create new objects automatically
in a local directory called B2BII_MC_database if they are missing. Also
all scripts in b2bii/examples are updated to take an extra parameter
"mc" or "data" to specify what we are running on.

Database payloads for real data are cached in a directory called
B2BII_database. This directory serves as a fallback in case connection
to the central database fails but it can be deleted.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@25560 b38c6c8f-bee8-465d-8448-9abe6915d496